added new page to list holidays
